requirements.txt
openai
pytube
ffmpeg-python
language-tool-python
deep-translator
gTTS
httpx==0.23.0
yt-dlp
textstat
pydub
transformers
torch
torchaudio
bark
whisper
srt
streamlit
demucs
azure-cognitiveservices-speech
praat-parselmouth
python-dotenv

